Fix warning with gcc-10.1.0: - order of initialization - signed/unsigned comparison - unhandled case in switch - object slicing in catch clause - unused variable - self initialization

+++ src/game/startup/network/client/networkclientgamereconnection.h
@@ -51,7 +51,7 @@ public:
 private:
 	cSignalConnectionManager signalConnectionManager;
 
-	size_t localPlayerNr;
+	int localPlayerNr;
 	std::vector<cPlayerBasicData> players;
 
 	std::shared_ptr<cStaticMap> staticMap;
